+++ Paws4Me/Paws4Me/Controllers/Logic/PetSingleDetailsViewController.swift
@@ -32,28 +32,3 @@ class PetSingleDetailsViewController: UIViewController {
         self.navigationController?.navigationBar.isHidden = false
     }
 }
-
-extension UIImage {
-    public static func loadFrom(url: URL, completion: @escaping (_ image: UIImage?) -> Void) {
-        DispatchQueue.global().async {
-            if let data = try? Data(contentsOf: url) {
-                DispatchQueue.main.async {
-                    completion(UIImage(data: data))
-                }
-            } else {
-                DispatchQueue.main.async {
-                    completion(nil)
-                }
-            }
-        }
-    }
-    public static func displayImgFromUrl(url: String, petImageView: UIImageView)
-    -> UIImageView {
-        guard let url = URL(string: url) else { return UIImageView() }
-        UIImage.loadFrom(url: url) {image in
-            petImageView.layer.cornerRadius = 10
-            petImageView.image = image
-        }
-        return petImageView
-    }
-}
