+++ Paws4Me/Paws4Me/Controllers/Logic/SignInViewController.swift
@@ -56,16 +56,3 @@ func addErrorBorderBoth(_ textField1: UITextField, _ textField2: UITextField,
     textField1.layer.addSublayer(bottomLine)
     textField2.layer.addSublayer(bottomLine2)
 }
-
-extension UIViewController {
-    func displayErrorAlert(alertTitle: String,
-                           alertMessage: String,
-                           alertActionTitle: String,
-                           alertDelegate: UIViewController) {
-        let alert = UIAlertController(title: alertTitle,
-                                      message: alertMessage,
-                                      preferredStyle: UIAlertController.Style.alert)
-        alert.addAction(UIAlertAction(title: alertActionTitle, style: UIAlertAction.Style.default, handler: nil))
-        alertDelegate.present(alert, animated: true, completion: nil)
-    }
-}
