+++ Paws4Me/Paws4Me/Controllers/Logic/ViewController.swift
@@ -155,11 +155,3 @@ class ViewController: UIViewController, UISearchBarDelegate {
            }
        }
    }
-
-   extension UIViewController {
-       func displayAlert(title: String, message: String, buttonTitle: String) {
-           let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertController.Style.alert)
-           alert.addAction(UIAlertAction(title: buttonTitle, style: UIAlertAction.Style.default, handler: nil))
-           self.present(alert, animated: true, completion: nil)
-       }
-   }
